@@ -18,16 +18,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: "^3.3.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  responsive_builder: ^0.2.0+2
+  responsive_builder: 0.7.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: 1.0.6
 
 dev_dependencies:
   flutter_test:
