@@ -4,5 +4,5 @@ class DataModel{
   final String date;
   final String time;
 
-  DataModel({this.serviceName, this.date, this.time});
+  DataModel({required this.serviceName, required this.date, required this.time});
 }
